requirements.txt
# Snowflake ML Jobs - Pyomo Optimization POC
# ==========================================

# Snowflake packages
snowflake-snowpark-python>=1.11.0
snowflake-ml-python>=1.9.2

# Optimization
pyomo>=6.7.0

# Solver - HiGHS is pip-installable and works in containerized environments
highspy>=1.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Development/testing
python-dotenv>=1.0.0

